fix os.path.split usage
os.path.split does not split a path at the dirsep. It does so only
for the last spearator. So you get a tuple with everything except
the last element, and the last element.
os.path.join in turn expects one path element and than a variable
number of arguments. Instead it got a list, which it interprets
as a single path element. Under windows, this triggered exceptions
if the path starts with a disk drive letter during import.