Major cleanup (such as PEP8 et al, see details in CHANGES.rst).
This is the non-pep8 part of commit
cc2c418371ce02849beae939f568ff01615dc3c3 on branch
jensens-pep8-and-cleanup

I copied the files over, so the result would be exactly the same.  But
now the boring pep8 cleanup is split off, so it is easier to focus on
the interesting part.